Extracted common IOApp signatures
