+++ core/js/src/main/scala/cats/effect/IOAppPlatform.scala
@@ -20,14 +20,12 @@ import scala.concurrent.CancellationException
 import scala.concurrent.duration._
 import scala.scalajs.js
 
-trait IOApp {
+private[effect] abstract class IOAppPlatform { this: IOApp =>
 
   private[this] var _runtime: unsafe.IORuntime = null
 
   protected def runtime: unsafe.IORuntime = _runtime
   protected def runtimeConfig: unsafe.IORuntimeConfig = unsafe.IORuntimeConfig()
-
-  def run(args: List[String]): IO[ExitCode]
 
   final def main(args: Array[String]): Unit = {
     if (runtime == null) {
@@ -88,13 +86,4 @@ trait IOApp {
     if (js.typeOf(js.Dynamic.global.process) != "undefined") {
       js.Dynamic.global.process.exitCode = code.code
     }
-}
-
-object IOApp {
-
-  trait Simple extends IOApp {
-    def run: IO[Unit]
-    final def run(args: List[String]): IO[ExitCode] = run.as(ExitCode.Success)
-  }
-
 }
