+++ core/jvm/src/main/scala/cats/effect/IOAppPlatform.scala
@@ -20,7 +20,7 @@ import scala.concurrent.{blocking, CancellationException}
 
 import java.util.concurrent.CountDownLatch
 
-trait IOApp {
+private[effect] abstract class IOAppPlatform { this: IOApp =>
 
   private[this] var _runtime: unsafe.IORuntime = null
   protected def runtime: unsafe.IORuntime = _runtime
@@ -29,8 +29,6 @@ trait IOApp {
 
   protected def computeWorkerThreadCount: Int =
     Math.max(2, Runtime.getRuntime().availableProcessors())
-
-  def run(args: List[String]): IO[ExitCode]
 
   final def main(args: Array[String]): Unit = {
     if (runtime == null) {
@@ -139,14 +137,4 @@ trait IOApp {
         Thread.currentThread().interrupt()
     }
   }
-
-}
-
-object IOApp {
-
-  trait Simple extends IOApp {
-    def run: IO[Unit]
-    final def run(args: List[String]): IO[ExitCode] = run.as(ExitCode.Success)
-  }
-
 }
